% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpcstats.R
\name{binning}
\alias{binning}
\alias{binning.tidyvpcobj}
\title{binning}
\usage{
binning(o, ...)

\method{binning}{tidyvpcobj}(
  o,
  bin,
  data = o$data,
  xbin = "xmedian",
  centers,
  breaks,
  nbins,
  altx,
  stratum = NULL,
  by.strata = T,
  ...
)
}
\arguments{
\item{o}{tidyvpc object}

\item{...}{Other arguments to include}

\item{bin}{Character string indicating binning method or unquoted variable name if binning on x-variable}

\item{data}{Observed data supplied in \code{observed()} function}

\item{xbin}{Character string indicating midpoint type for binning}

\item{centers}{Numeric vector of centers for binning. Use \code{bin = "centers"} if supplying centers}

\item{breaks}{Numeric vector of breaks for binning. Use \code{bin = "breaks"} if supplying breaks}

\item{nbins}{Numeric number indicating the number of bins to use}

\item{altx}{Unquoted variable name in observed data for elternative x-variable binning}

\item{stratum}{List indicating the name of stratification variable and level if using different binning methods by strata}

\item{by.strata}{Logical indicating whether binning should be performed by strata}
}
\description{
Binning methods for Visual Predictive Check (VPC)
}
\details{
This function executes binning methods available in classInt i.e. "jenks", "kmeans", "sd", "pretty", "pam", etc.
You may also bin directly on "x-variable" or alternatively specify "centers" or "breaks".
}
\examples{
\dontrun{
 # Binning on x-variable
vpc <- observed(exampleobs, x=TIME, y=DV) \%>\%
    simulated(examplesim, y=DV) \%>\%
    binning(bin = TIME) \%>\%
    vpcstats()
    
 # Binning using ntile and xmean for midpoint
vpc <- observed(exampleobs, x=TIME, y=DV) \%>\%
    simulated(examplesim, y=DV) \%>\%
    binning(bin = "ntile", nbins = 8, xbin = "xmean") \%>\%
    vpcstats()
    
 # Binning using centers
vpc <- observed(exampleobs, x=TIME, y=DV) \%>\%
    simulated(examplesim, y=DV) \%>\%
    binning(bin = "centers", centers = c(1,3,5,7)) \%>\%
    vpcstats()
        
 # Different Binning for each level of Strata
vpc <- observed(exampleobs, x=TIME, y=DV) \%>\%
    simulated(examplesim, y=DV) \%>\%
    stratify(~ ISM) \%>\%
    binning(stratum = list(ISM = 0), bin = "pam", nbins = 9, by.strata = T) \%>\%
    binning(stratum = list(ISM = 1), bin = "breaks", breaks = c(2,5,7,10), by.strata = T) \%>\%
    vpcstats()
}
}
\seealso{
\code{\link{observed}} \code{\link{simulated}} \code{\link{censoring}} \code{\link{predcorrect}} \code{\link{stratify}} \code{\link{binless}} \code{\link{vpcstats}}
}
