% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleobs}
\alias{exampleobs}
\title{Example observed data from vpc package.}
\format{A data frame with 600 rows and 7 variables:
\describe{
  \item{ID}{Subect identifier}
  \item{TIME}{Time}
  \item{DV}{Concentration of drug}
  \item{AMT}{Amount of dosage initially administered at DV = 0, TIME = 0}
  \item{DOSE}{Dosage amount}
  \item{MDV}{Dummy indiciating missing dependent variable value}
  \item{ISM}{Dummy variable indicating subject's gender (ISM = 0, ISM = 1)}
}}
\source{
\code{\link[vpc]{simple_data}}
}
\usage{
exampleobs
}
\description{
An observed dataset from a hypothetical PK model.
}
\keyword{datasets}
