% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpcstats.R
\name{nopredcorrect}
\alias{nopredcorrect}
\alias{nopredcorrect.tidyvpcobj}
\title{Remove prediction correction for Visual Predictive Check (VPC)}
\usage{
nopredcorrect(o, ...)

\method{nopredcorrect}{tidyvpcobj}(o, ...)
}
\arguments{
\item{o}{A \code{tidyvpcobj}.}

\item{...}{Other arguments to include.}
}
\description{
Optional function to use indicating no pred correction for VPC.
}
