% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcVPC.R
\name{norefcorrect}
\alias{norefcorrect}
\alias{norefcorrect.tidyvpcobj}
\title{norefcorrect}
\usage{
norefcorrect(o, ...)

\method{norefcorrect}{tidyvpcobj}(o, ...)
}
\arguments{
\item{o}{tidyvpcobj}

\item{...}{other arguments to include}
}
\description{
No ref correction for Visual Predictive Check (VPC)
}
\details{
Optional function to use indicating no ref correction for VPC.
}
