% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcVPC.R
\name{norefcorrectCFB}
\alias{norefcorrectCFB}
\alias{norefcorrectCFB.tidyvpcobj}
\title{norefcorrectCFB}
\usage{
norefcorrectCFB(o, ...)

\method{norefcorrectCFB}{tidyvpcobj}(o, ...)
}
\arguments{
\item{o}{tidyvpcobj}

\item{...}{other arguments to include}
}
\description{
No ref correction for Visual Predictive Check (VPC) for the change from baseline.
}
\details{
Optional function to use indicating no ref correction for the change from baseline.
}
