% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{obs_cat_data}
\alias{obs_cat_data}
\title{Example observed data with categorical DV}
\format{
A data frame with 4014 rows and 4 variables:
\describe{
  \item{PID_code}{Subject identifier}
  \item{agemonths}{Time}
  \item{zlencat}{Categorical DV with the 3 levels}
  \item{Country_ID_code}{Country code for stratification}
}
}
\source{
Certara University
}
\usage{
obs_cat_data
}
\description{
An observed dataset with 3 levels of categorical DV.
}
\keyword{datasets}
