% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpcstats.R
\name{plotbinless.vpcstatsobj}
\alias{plotbinless.vpcstatsobj}
\title{Plot a binless vpcstats object}
\usage{
plotbinless.vpcstatsobj(
  x,
  ...,
  show.points = TRUE,
  xlab = NULL,
  ylab = NULL,
  color = c("red", "blue", "red"),
  linetype = c("dashed", "solid", "dashed"),
  legend.position = "top",
  facet.scales = "free"
)
}
\arguments{
\item{x}{A vpcstats object.}

\item{...}{Other arguments to include.}

\item{show.points}{If TRUE (default) returns observed data points in plot.}

\item{xlab}{X label text to include}

\item{ylab}{y label text to include}

\item{color}{line and fill color for plot}

\item{linetype}{line type}

\item{legend.position}{legend orientation in plot}

\item{facet.scales}{facet formatting}
}
\value{
Returns binless vpc plot.
}
\description{
Plot a binless vpcstats object
}
