% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpcstats.R
\name{predcorrect}
\alias{predcorrect}
\alias{predcorrect.tidyvpcobj}
\title{predcorrect}
\usage{
predcorrect(o, ...)

\method{predcorrect}{tidyvpcobj}(o, pred, data = o$data, ..., log = FALSE)
}
\arguments{
\item{o}{tidyvpc object}

\item{...}{Other arguments to include}

\item{pred}{prediction variable in observed data}

\item{data}{observed data supplied in \code{observed()} function}

\item{log}{logical indicating whether DV was modeled in logarithimic scale}
}
\description{
Prediction corrected Visual Predictive Check (pcVPC)
}
\details{
Specify prediction variable for pcVPC
}
\seealso{
\code{\link{observed}} \code{\link{simulated}} \code{\link{censoring}} \code{\link{stratify}} \code{\link{binning}} \code{\link{binless}} \code{\link{vpcstats}}
}
