% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpcstats.R
\name{print.tidyvpcobj}
\alias{print.tidyvpcobj}
\title{Print a \code{tidyvpcobj}}
\usage{
\method{print}{tidyvpcobj}(x, ...)
}
\arguments{
\item{x}{An \code{tidyvpcobj}.}

\item{...}{Further arguments can be specified but are ignored.}
}
\value{
Returns \code{x} invisibly.
}
\description{
Print generic used to return information about VPC.
}
