% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcVPC.R
\name{refcorrect}
\alias{refcorrect}
\alias{refcorrect.tidyvpcobj}
\title{refcorrect}
\usage{
refcorrect(o, ...)

\method{refcorrect}{tidyvpcobj}(
  o,
  yref,
  xref,
  ypred,
  refdata,
  obsdata = o$data,
  simdata = o$simdata,
  PRED_simdata = NULL,
  PRED_refdata = NULL,
  IDV_refcorr = FALSE,
  log = o$lnDV,
  ...
)
}
\arguments{
\item{o}{tidyvpc object}

\item{...}{Other arguments to include}

\item{yref}{dependent variable}

\item{xref}{independent variable}

\item{ypred}{population prediction variable}

\item{refdata}{simulated reference data}

\item{PRED_simdata}{the simulated data with the prediction variable of the independent variable, with the same variable name as xref}

\item{PRED_refdata}{the simulated reference data with the prediction variable of the independent variable, with the same variable name as xref}

\item{IDV_refcorr}{flag if the Independent variable should be reference corrected as well}
}
\value{
Updates \code{tidyvpcobj} with required information to performing Reference correction which include \code{cor} logical indicating whether
  corrected VPC is to be performed, \code{cor.log} logical indicating whether the DV is on a log-scale.
}
\description{
reference corrected Visual Predictive Check (rcVPC)
}
\details{
This function creates a reference corrected Visual Predictive Check (rcVPC)
}
\examples{
require(magrittr)

obs_data <- obs_data[MDV == 0]
sim_data <- sim_data[MDV == 0]
ref_data <- ref_data[MDV == 0]

  vpc <- observed(obs_data, x=TIME, y=DV) \%>\%
       simulated(sim_data, y=DV) \%>\%
       refcorrect(refdata = ref_data, yref=DV, xref=TIME, ypred=PRED) \%>\% 
       binning(bin = NTIME) \%>\%
       vpcstats()
       
}
\seealso{
\code{\link{observed}} \code{\link{simulated}} \code{\link{censoring}} \code{\link{binning}} \code{\link{vpcstats}}
}
