% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcVPC.R
\name{refcorrectCFB}
\alias{refcorrectCFB}
\alias{refcorrectCFB.tidyvpcobj}
\title{refcorrectCFB}
\usage{
refcorrectCFB(o, ...)

\method{refcorrectCFB}{tidyvpcobj}(
  o,
  yref,
  xref,
  refdata,
  obsdata = o$data,
  simdata = o$simdata,
  PRED_simdata = NULL,
  PRED_refdata = NULL,
  lb = -1,
  IDV_refcorr = FALSE,
  log = o$lnDV,
  ...
)
}
\arguments{
\item{o}{tidyvpc object}

\item{...}{Other arguments to include}

\item{yref}{variable name for the change from baseline data item, it should be the same name across the following three input data}

\item{xref}{independent variable}

\item{refdata}{the simulated reference data}

\item{PRED_simdata}{change from baseline prediction for the simulated data}

\item{PRED_refdata}{change from baseline prediction for the simulated reference data}

\item{lb}{lower bound for the change from baseline, e.g., -100.}

\item{IDV_refcorr}{flag if the Independent variable should be reference corrected as well}
}
\value{
Updates \code{tidyvpcobj} with required information to performing Reference correction which include \code{cor} logical indicating whether
  corrected VPC is to be performed.
}
\description{
reference corrected Visual Predictive Check (rcVPC) for the change from baseline
}
\details{
This function creates a reference corrected Visual Predictive Check (rcVPC) for the change from baseline.
}
\examples{
require(magrittr)

obs_data    <- obs_data[MDV == 0]
sim_data    <- sim_data[MDV == 0]
PCFB_sdata  <- PCFB_sim_data[MDV == 0]
ref_data    <- ref_data[MDV == 0]
PCFB_rdata  <- PCFB_ref_data[MDV == 0]

rcVPC <- observed(odata, x = CP, yobs = CDV) \%>\% 
 simulated(sdata, ysim = CDV) \%>\%
 refcorrectCFB(lb           = -1,
               xref         = CP,
               yref         = CDV,
               PRED_simdata = PCFB_sdata,
               refdata      = ref_data,
               PRED_refdata = PCFB_rdata) \%>\% 
 binning(bin = "centers", centers = c(-Inf,0.01,25,50,75,100,150,Inf)) \%>\%
 vpcstats(qpred = c(0.1, 0.5, 0.9))
       
}
\seealso{
\code{\link{observed}} \code{\link{simulated}} \code{\link{censoring}} \code{\link{binning}} \code{\link{vpcstats}}
}
