% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpcstats.R
\name{runShinyVPC}
\alias{runShinyVPC}
\title{runShinyVPC}
\usage{
runShinyVPC()
}
\description{
Run Shiny app for tidyvpc
}
\details{
Use this function to run Shiny application to parameterize VPC from a GUI and generate corresponding tidyvpc code to derive VPC.
}
\seealso{
\href{https://github.com/jameswcraig/shiny-vpc/blob/master/README.md/}{Shiny-VPC GitHub}
}
