% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_cat_data}
\alias{sim_cat_data}
\title{Example simulated data with categorical DV}
\format{
A data frame with 401400 rows and 4 variables:
\describe{
  \item{PID_code}{Subject identifier}
  \item{IVAR}{Time}
  \item{DV}{Categorical DV with 3 levels}
  \item{Replicate}{Replicate num for simulation}
}
}
\source{
Certara University
}
\usage{
sim_cat_data
}
\description{
A simulated dataset with the 3 levels of categorical DV across 100 replicates.
}
\keyword{datasets}
