% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_data}
\alias{sim_data}
\title{Example simulated data from vpc package.}
\format{A data frame with 60000 rows and 10 variables:
\describe{
  \item{ID}{Subect identifier}
  \item{REP}{Replicate num for simulation}
  \item{TIME}{Time}
  \item{DV}{Concentration of drug}
  \item{IPRED}{Individual prediction variable}
  \item{PRED}{Population prediction variable}
  \item{AMT}{Amount of dosage initially administered at DV = 0, TIME = 0}
  \item{DOSE}{Dosage amount}
  \item{MDV}{Dummy indiciating missing dependent variable value}
  \item{NTIME}{Nominal Time}
}}
\source{
\code{\link[vpc]{simple_data}}
}
\usage{
sim_data
}
\description{
A simulated dataset from a hypothetical PK model with 100 replicates.
}
\keyword{datasets}
