% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpcstats.R
\name{simulated}
\alias{simulated}
\alias{simulated.tidyvpcobj}
\title{observed}
\usage{
simulated(o, ...)

\method{simulated}{tidyvpcobj}(o, data, ysim, ...)
}
\arguments{
\item{o}{tidyvpcobj}

\item{...}{other arguments}

\item{data}{data.frame or data.table of simulated data}

\item{ysim}{numeric y-variable, typically named DV}
}
\description{
Specify simulated dataset and variables for VPC
}
\details{
The simulated function is the second function in the vpc piping chain and is used for specifying simulated data and variables for VPC
}
