% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpcstats.R
\name{vpcstats}
\alias{vpcstats}
\alias{vpcstats.tidyvpcobj}
\title{vpcstats}
\usage{
vpcstats(o, ...)

\method{vpcstats}{tidyvpcobj}(
  o,
  qpred = c(0.05, 0.5, 0.95),
  ...,
  conf.level = 0.95,
  quantile.type = 7
)
}
\arguments{
\item{o}{tidyvpc object}

\item{...}{Other arguments to include}

\item{qpred}{Numeric vector of length 3 specifying quantile prediction interval}

\item{conf.level}{Numeric specifying confidence level}

\item{quantile.type}{Numeric indicating quantile type. See \code{\link[stats]{quantile}}}
}
\description{
Compute VPC statistics
}
\details{
Compute predictional interval statistics for VPC
}
\seealso{
\code{\link{observed}} \code{\link{simulated}} \code{\link{censoring}} \code{\link{stratify}} \code{\link{binning}} \code{\link{binless}} \code{\link{predcorrect}}
}
