% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpcstats.R
\name{binless}
\alias{binless}
\alias{binless.tidyvpcobj}
\title{binless}
\usage{
binless(o, ...)

\method{binless}{tidyvpcobj}(
  o,
  qpred = c(0.05, 0.5, 0.95),
  optimize = TRUE,
  optimization.interval = c(0, 7),
  conf.level = 0.95,
  loess.ypc = FALSE,
  lambda = NULL,
  span = NULL,
  ...
)
}
\arguments{
\item{o}{tidyvpc object}

\item{...}{other arguments}

\item{qpred}{numeric vector of length 3 specifying quantiles (lower, median, upper) i.e. \code{c(0.1, 0.5, 0.9)}}

\item{optimize}{logical indicating whether lambda and span should be optimized using AIC}

\item{optimization.interval}{numeric vector of length 2 specifying interval for lambda optimization}

\item{conf.level}{numeric confidence level for binless fit}

\item{loess.ypc}{logical indicating loess precition corrected. Must first use \code{predcorrect()} if \code{loess.ypc = TRUE}}

\item{lambda}{numeric vector of length 3 specifying lambda values for each quantile}

\item{span}{numeric number between 0,1 specying smoothing paramter for loess prediction corrected}
}
\description{
Perform binless Visual Predictive Check (VPC)
}
\details{
Use this function in subsitute of traditional binning methods to derive VPC using additive quantile regression and loess for pcVPC.
}
\examples{
\dontrun{
Binless example with user specified lambda values stratified on "ISM" with 2 levels (0, 1)
 
 exampleobs$SEX <- rep(c("F", "M"), len=nrow(exampleobs))
 lambda_strat <- data.table(
 ISM0 = c(3,5,2),
 ISM1 = c(1, 3, 4),
 )
 
 vpc <- observed(dataObs, y = DV, x = TIME) \%>\%
      simulated(dataSim, y = DV) \%>\%
      stratify(~ ISM) \%>\%
      binless(qpred = c(0.1, 0.5, 0.9), optimize = FALSE, lambda = lambda_strat) \%>\%
      vpcstats()
      
plot(vpc)
}
}
