% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpcstats.R
\name{censoring}
\alias{censoring}
\alias{censoring.tidyvpcobj}
\title{censoring}
\usage{
censoring(o, ...)

\method{censoring}{tidyvpcobj}(o, blq, lloq, data = o$data, ...)
}
\arguments{
\item{o}{tidyvpc object}

\item{...}{Other arguments to include}

\item{blq}{blq variable if present in observed data}

\item{lloq}{lloq variable if present in observed data. Use numeric to specify lloq value}

\item{data}{observed data supplied in \code{observed()} function}
}
\description{
Censoring observed data for Visual Predictive Check (VPC)
}
\details{
Specify censoring variables or censoring value for VPC using this function
}
