% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpcstats.R
\name{generics}
\alias{generics}
\alias{vpcstats}
\title{Perform a Visual Predictive Check (VPC) computation}
\usage{
vpcstats(o, ...)
}
\arguments{
\item{o}{tidyvpc object}

\item{...}{Other arguments to include}

\item{qpred}{Numeric vector of length 3 specifying quantile prediction interval}

\item{conf.level}{Numeric specifying confidence level}

\item{quantile.type}{Numeric indicating quantile type. See \code{\link[stats]{quantile}}}
}
\description{
These functions work together to calculate the statistics that are plotted
in a VPC. They would typically be chained together using the "pipe" operator
(see Examples).

Compute VPC statistics
}
\details{
Compute predictional interval statistics for VPC
}
\examples{

\dontrun{
library(data.table)
library(quantreg)
library(magrittr)
library(vpc)

exampleobs <- as.data.table(vpc::simple_data$obs)[MDV == 0]
examplesim <- as.data.table(vpc::simple_data$sim)[MDV == 0]
exampleobs$PRED <- examplesim[REP == 1, PRED]
exampleobs$SEX <- rep(c("F", "M"), len=nrow(exampleobs))

vpc <- observed(exampleobs, x=TIME, y=DV) \%>\%
    simulated(examplesim, y=DV) \%>\%
    stratify(~ ISM) \%>\%
    binning(TIME) \%>\%
    predcorrect(pred=PRED) \%>\%
    vpcstats()

plot(vpc)

# Example with 2-way stratification

vpc <- vpc \%>\%
    stratify(SEX ~ ISM) \%>\%
    binning(TIME) \%>\%
    predcorrect(pred=PRED) \%>\%
    vpcstats()

plot(vpc)
}

}
\seealso{
\code{\link{observed}} \code{\link{simulated}} \code{\link{censoring}} \code{\link{stratify}} \code{\link{binning}} \code{\link{binless}} \code{\link{predcorrect}}
}
