% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpcstats.R
\name{generics}
\alias{generics}
\title{Perform a Visual Predictive Check (VPC) computation}
\arguments{
\item{o}{An object.}

\item{...}{Additional arguments.}
}
\description{
These functions work together to calculate the statistics that are plotted
in a VPC. They would typically be chained together using the "pipe" operator
(see Examples).
}
