% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpcstats.R
\name{nopredcorrect}
\alias{nopredcorrect}
\alias{nopredcorrect.tidyvpcobj}
\title{nopredcorrect}
\usage{
nopredcorrect(o, ...)

\method{nopredcorrect}{tidyvpcobj}(o, ...)
}
\arguments{
\item{o}{tidyvpcobj}

\item{...}{other arguments to include}
}
\description{
No pred correction for Visual Predictive Check (VPC)
}
\details{
Optional function to use indicating no pred correction for VPC.
}
