% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpcstats.R
\name{simulated}
\alias{simulated}
\alias{simulated.tidyvpcobj}
\title{Specify simulated dataset and variables for VPC}
\usage{
simulated(o, ...)

\method{simulated}{tidyvpcobj}(o, data, ysim, ..., xsim)
}
\arguments{
\item{o}{A \code{tidyvpcobj}.}

\item{...}{Other arguments.}

\item{data}{A \code{data.frame} of simulated data.}

\item{ysim}{Numeric y-variable, typically named DV.}

\item{xsim}{Numeric x-variable, typically named TIME.}
}
\value{
A \code{tidyvpcobj} containing simulated dataset \code{sim} formatted with columns \code{x}, \code{y}, and \code{repl}, which indicates the replicate number.
 The column \code{x} is used from the \code{observed()} function. Resulting dataset is of class \code{data.frame} and \code{data.table}.
}
\description{
The simulated function is used for specifying simulated input data and
variables for VPC. Note: Simulated data must not contain missing DV and may
require filtering \code{MDV == 0} before generating VPC. Simulated data must
be ordered by: Replicate, Subject (ID), IVAR (Time).
}
\examples{
require(magrittr)

vpc <- observed(obs_data, x=TIME, y=DV) \%>\%
    simulated(sim_data, y=DV)

}
\seealso{
\code{\link{observed}} \code{\link{censoring}} \code{\link{stratify}} \code{\link{predcorrect}} \code{\link{binning}} \code{\link{binless}} \code{\link{vpcstats}}
}
