% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpcstats.R
\name{stratify}
\alias{stratify}
\alias{stratify.tidyvpcobj}
\title{stratify}
\usage{
stratify(o, ...)

\method{stratify}{tidyvpcobj}(o, formula, data = o$data, ...)
}
\arguments{
\item{o}{tidyvpc object}

\item{...}{Other arguments to include}

\item{formula}{formula for stratfication}

\item{data}{Observed data supplied in \code{observed()} function}
}
\description{
Stratification for Visual Predictive Check (VPC)
}
\details{
specify stratification variables for VPC using this function
}
\seealso{
\code{\link{observed}} \code{\link{simulated}} \code{\link{censoring}} \code{\link{predcorrect}} \code{\link{binning}} \code{\link{binless}} \code{\link{vpcstats}}
}
